fix(component): require cycles
