+++ src/components/social/profile/profileActivityList3.component.tsx
@@ -1,7 +1,10 @@
 import React from 'react';
 import {
+  Dimensions,
+  FlexStyle,
   ImageSourcePropType,
   ListRenderItemInfo,
+  ScaledSize,
 } from 'react-native';
 import {
   StyleType,
@@ -14,9 +17,9 @@ import {
   ListProps,
 } from '@kitten/ui';
 import {
-  ProfileActivityList2Item,
-  ProfileActivityList2ItemProps,
-} from './profileActivityList2Item.component';
+  ProfileActivityList3Item,
+  ProfileActivityList3ItemProps,
+} from './profileActivityList3Item.component';
 
 // @ts-ignore (override `renderItem` prop)
 interface ComponentProps extends ListProps {
@@ -25,22 +28,34 @@ interface ComponentProps extends ListProps {
   renderItem?: (info: ListRenderItemInfo<ImageSourcePropType>, style: StyleType) => React.ReactElement<any>;
 }
 
-type ListItemElement = React.ReactElement<ProfileActivityList2ItemProps>;
+type ListItemElement = React.ReactElement<ProfileActivityList3ItemProps>;
 
-export type ProfileActivityList2Props = ThemedComponentProps & ComponentProps;
+const dimensions: ScaledSize = Dimensions.get('window');
 
-class ProfileActivityList2Component extends React.Component<ProfileActivityList2Props> {
+export type ProfileActivityList3Props = ThemedComponentProps & ComponentProps;
+
+class ProfileActivityList3Component extends React.Component<ProfileActivityList3Props> {
+
+  static defaultProps: Partial<ProfileActivityList3Props> = {
+    numColumns: 3,
+  };
 
   private onItemPress = (index: number) => {
     this.props.onItemPress(index);
   };
 
   private renderListItemElement = (item: ImageSourcePropType): ListItemElement => {
-    const { themedStyle } = this.props;
+    const { themedStyle, numColumns } = this.props;
+
+    const size: number = dimensions.width / numColumns;
+    const sizeStyle: FlexStyle = {
+      width: size,
+      height: size,
+    };
 
     return (
-      <ProfileActivityList2Item
-        style={themedStyle.item}
+      <ProfileActivityList3Item
+        style={[themedStyle.item, sizeStyle]}
         activeOpacity={0.75}
         source={item}
         onPress={this.onItemPress}
@@ -61,8 +76,6 @@ class ProfileActivityList2Component extends React.Component<ProfileActivityList2
 
     return (
       <List
-        horizontal={true}
-        showsHorizontalScrollIndicator={false}
         {...restProps}
         contentContainerStyle={[themedStyle.container, contentContainerStyle]}
         renderItem={this.renderItem}
@@ -71,12 +84,9 @@ class ProfileActivityList2Component extends React.Component<ProfileActivityList2
   }
 }
 
-export const ProfileActivityList2 = withStyles(ProfileActivityList2Component, (theme: ThemeType) => ({
+export const ProfileActivityList3 = withStyles(ProfileActivityList3Component, (theme: ThemeType) => ({
   container: {},
   item: {
-    width: 120,
-    height: 120,
-    marginHorizontal: 8,
-    backgroundColor: theme['color-basic-200'],
+    backgroundColor: theme['color-basic-100'],
   },
 }));
