+++ src/components/social/profile/profileSocials.component.tsx
@@ -12,7 +12,6 @@ import {
   withStyles,
 } from '@kitten/theme';
 import { Text } from '@src/components/common';
-import { Button } from '@kitten/ui';
 
 interface ComponentProps {
   followers: React.ReactText;
